+++ modules/pipewire/default.nix
@@ -1,9 +1,12 @@
-_: {
+{ self, pkgs, ... }: {
   services.pipewire = {
     enable = true;
     alsa.enable = true;
     alsa.support32Bit = true;
     pulse.enable = true;
+
+    wireplumber.configPackages = with self.packages.${pkgs.system};
+      [ wireplumber-disable-pci-0 ];
   };
 
   security.rtkit.enable = true;